refactor(chat): extract provider auth and model arg parsing helpers
